requirements.txt
jax==0.5.0
matplotlib==3.10.0
pandas==2.2.3
PyYAML==6.0.2